# spec/openapi.yaml
openapi: '3.1.0'
info:
  title: Sherpa Web API
  version: '0.1.1'
  description: |
    Sherpa API endpoints
servers:
  - description: Sherpa Production Server
    url: https://sherpa.gruene.de/sherpa/ws/m2m/
...
security:
  - mutualTLS: []
paths:
...
  /saml/party/newusers:
    post:
      operationId: SamlFindNewUsersByDate
      summary: Find new users by date
      description: Find created users by date. Response does not include memberships or roles.
      tags:
        - saml
      requestBody:
        required: true
        description: Find new user params
        content:
          application/json:
            schema:
              type: object
              properties:
                from:
                  type: string
                  format: date-time
                  example: 2021-01-27T16:50:37+0100
                to:
                  type: string
                  format: date-time
                  example: 2021-01-28T16:50:37+0100
              required:
                - from
                - to
      responses:
        '200':
          description: List of new users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SamlNewUser'
        '415':
          description: Missing Request Body
        '500':
          $ref: '#/components/responses/SamlServerError'
  /saml/party/list:
    post:
      operationId: SamlGetUsersById
      summary: Get users by ids
      description: Get detailed information about users and their memberships and roles by id.
      tags:
        - saml
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                partyIdList:
                  type: array
                  items:
                    type: string
                  example: ['12345678', '12345679']
              required:
                - partyIdList
      responses:
        '200':
          description: List of new users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SamlUser'
        '500':
          $ref: '#/components/responses/SamlServerError'
  /saml/v1/users:
    get:
      operationId: SamlFindUsers
      summary: Find users
      description: Find and filter users. Get detailed information about the user memberships and roles.
      tags:
        - saml
      parameters:
        - name: limit
          in: query
          required: false
          description: Max amount of users to return
          schema:
            type: integer
            maximum: 1000
            minimum: 1
            default: 200
        - name: after
          in: query
          required: false
          description: Last user id, used for keyset pagination.
          schema:
            type: string
            example: '1234567'
        - name: modified_after
          in: query
          required: false
          description: ISO8601 compliant date. Only list users modified after given date.
          schema:
            type: string
            format: date-time
            example: 2023-01-01T00:00:00+0100
        - name: modified_before
          in: query
          required: false
          description: ISO8601 compliant date. Only list users modified before given date.
          schema:
            type: string
            format: date-time
            example: 2023-02-01T00:00:00+0100
        - name: include[]
          in: query
          required: false
          description: Filter users by user ids
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Find user response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SamlFindUsersResponse'

  # gnetz endpoints
...
components:
...
  responses:
...
    SamlServerError:
      description: Generic error response for authentication, validation and others
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SamlError'

    # gnetz responses
...
  schemas:
...
    DivisionKey:
      type: string
      description: >
        External Division identifier.
          - First digit is the index of the hierarchy
            1. GR
            2. GJ
            3. KPV
          - second and third digit is the *LV*
          - next three is the *KV*
          - last two is the *OV*
        All undefined hierarchy levels are nulled.
      example: '10000000'
      # examples: # from OAS 3.1 on
      #   warstein:
      #     summary: OV Warstein
      #     description: GR - NRW - SO - Warstein
      #     value: "11004613"
      #   bv:
      #     summary: Bundesverband GJ
      #     description: GJ - null - null - null - null
      #     value: "20000000"
      minLength: 8
      maxLength: 8
      pattern: '^[1-9][0-9]{7}$'
...
    SamlFindUsersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SamlUser'
        meta:
          type: object
          properties:
            hasNext:
              type: boolean
              description: Whether a next page exists or not
      required:
        - data
        - meta
    SamlNewUser:
      type: object
      description: new user
      properties:
        toType:
          type: string
          example: SamlOnboardingPartyTO
        id:
          type: string
          description: sherpa user id
          example: '12345678'
        name1:
          type: string
          description: Firstname
          example: John
        name2:
          type: string
          description: Lastname
          example: Doe
        email:
          type: [string, 'null']
          format: email
          example: john.doe@example.com
      required:
        - toType
        - id
        - name1
        - name2
        - email
    SamlUser:
      type: object
      properties:
        toType:
          type: string
          example: SamlSynchronizationPartyTO
        id:
          type: string
          description: sherpa user id
          example: '12345678'
        ownerIdentifier:
          $ref: '#/components/schemas/DivisionKey'
        ownerId:
          example: '11002607'
          type: string
        createdOn:
          type: string
          format: date-time
          example: 2020-07-22T13:38:30+0200
        lastModifiedOn:
          type: string
          format: date-time
          example: 2022-11-25T16:45:24+0100
        name1:
          type: string
          example: John
        name2:
          type: string
          example: Doe
        email:
          type: [string, 'null']
          format: email
          example: john.doe@example.com
        blockingNotice:
          description: Sperrvermerk in sherpa. Flag to indicate that user should not receive any emails.
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/SamlUserRole'
        memberships:
          type: array
          items:
            $ref: '#/components/schemas/SamlUserMembership'
      required:
        - toType
        - id
        - ownerIdentifier
        - ownerId
        - createdOn
        - lastModifiedOn
        - name1
        - name2
        - email
        - roles
        - memberships
    SamlUserRole:
      type: object
      properties:
        toType:
          type: string
          example: SamlSynchronizationRoleAssignmentTO
        id:
          type: string
          example: '13926537'
        ownerIdentifier:
          $ref: '#/components/schemas/DivisionKey'
        roleId:
          type: string
          example: '7555235'
        delegatedByOrganizationIdentifier:
          type: [string, 'null']
          description: can be a division key or an id for a non regular chapter starting with 5
          example: '512345'
        tags:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/SamlUserRoleTag'
      required:
        - toType
        - id
        - ownerIdentifier
        - roleId
        - delegatedByOrganizationIdentifier
        - tags
    SamlUserRoleTag:
      type: object
      properties:
        toType:
          type: string
          example: SamlGenericTagTO
        id:
          type: string
          example: '6592657'
        tagId:
          type: string
          example: '6590132'
        name:
          type: string
          example: eShop Rolle
      required:
        - toType
        - id
        - tagId
        - name
    SamlUserMembership:
      type: object
      properties:
        toType:
          type: string
          example: SamlSynchronizationMembershipTO
        id:
          type: string
        ownerIdentifier:
          $ref: '#/components/schemas/DivisionKey'
        memberOfStructureIdentifier:
          $ref: '#/components/schemas/DivisionKey'
      required:
        - toType
        - id
        - ownerIdentifier
        - memberOfStructureIdentifier
    SamlError:
      type: object
      properties:
        toType:
          type: string
          example: ExceptionTO
        exceptionClassName:
          type: string
          example: com.proximity.gp.common.exception.InvalidCredentialsException
        originalMessage:
          type: string
          example: api endpoint needs permission for SAML, caller got only GNETZ2
        httpStatusCode:
          type: string
          example: '500'
        userMessage:
          type: string
          example: Bei der Verarbeitung der Anfrage ist ein Fehler aufgetreten.
        stackTrace:
          type: string
        serverInfo:
          type: string
          example: Live-System (Version 1.4-89, Serverzeit 2023-04-09 09:21:47.519+0200)
        requestUuid:
          type: string
        source:
          type: string
          example: server
        type:
          type: string
          example: INVALID_CREDENTIALS
        code:
          type: string

    # gnetz schemas
...
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: >-
        The API authenticates callers solely by client certificate (mutual
        TLS). The certificate must be signed by the `Sherpa API CA` and carries
        the caller's permissions. mTLS is established at the transport layer, so
        the generated client does not handle it — configure your HTTP client
        (e.g. an axios `httpsAgent`) with the certificate and key.